# Reduce runtime work