I wanted to remove the need for locking mechanisms and `init` function as all data provided by this library is available at compile time. So I essentially rewrote this library to use `'static` data instead.

# Build Time
Here are compile time and binary size characteristics of my changes:

|Build|Old Time|New Time|Slope Ratio|Old Size*|New Size*|Size Change*|
|:-|:-:|:-:|-:|:-:|:-:|-:|
|debug; cold|12.05s|7.24s|-57.21%|19M|6.7M|-65.63%|
|debug; hot|1.11s|2.06s|-|-|-|-|
|release; cold|26.10s|6.15s|-86.84%|4.9M|6.6M|+34.95%|
|release; hot|23.55s|3.85s|-|-|-|-|

<sup>hot - `touch` only project files, dependencies already compiled</sup>
<sup>* - stripped rlib with data for 25 years</sup>

![build_times](https://github.com/user-attachments/assets/0d34c988-00d3-4acf-8d14-077812a09632)

Measured on _Intel i7-9750HF 2.60GHz_. In both cases I disabled all optional features, enabled all languages and used 2000 - 2025 data.</br>
As you can see from the picture above, slope of build time increase of old code was quite steeper than this new approach with tables.

|Build|Est. build time increase per holiday *|
|-:|:-:|
|Old - Debug|0.2641ms|
|Old - Release|0.6768ms|
|New - Debug|0.113ms|
|New - Release|0.0891ms|

<sup>* - 1 year of 101 countries = ~1400 entries</sup>

Diff is large because I added `holidays.csv` to the root in place of `crate::data::*` files, ignoring generated content in both old and new code, actual diff size is: +1408, -5668 (+/- 100LoC)

Now, onto the changes:

# Changes

## Codegen

Instead of generating code from a python script, I only made python dump data into a CSV file and `build.rs` use that file to generate code on the fly. This is preferred approach for these kinds of use cases. I'm also using the `phf` crate to generate a Map with `O(1)` lookup (for 1 country & 1 date query) - previously it was `O((log n)^3)`.<br/>
As the `phf::Map` is now completely static, this also avoids the need for `RwLock` around `DATA`.

Essentially, codegen now generates data instead of code that's executed at runtime - making everything run quicker.

Generated code is placed in `OUT_DIR` so that it doesn't pollute the repo.

`gen.py` now also produces `holidays.csv.hash` which contains the hash of used data in order to allow future automation with CI.

`build.rs` generates several jump tables to speed up different kinds of lookups.

## API

Instead of making a huge heap allocation with `init`, data returned from the library is baked into the binary (i.e. `&'static`). This makes everything `Copy`.

I removed the `Builder` as it's no longer needed.

I replaced existing iterator logic with 3 new iterators that handle different queries best - e.g. only country specified vs. single date and country vs. date range and country. This means that based on query parameters, best querying method will always be selected. These iterators also **don't clone data** or need to lock it, like the previous code did, which means they're much faster. They are exposed as a single opaque `Iter` struct that hides internals which the user likely won't care about.

Renamed `get` to `get_holidays` and `contains` to `is_holiday` so the purpose is clearer from the name.

`holidays::get_holidays` now returns that iterator, it was a _bug_ that it returned a single value as there could be multiple holidays occurring on the same day in a single country.

Both `get_holidays` and `is_holiday` support `None`, single or iterators/ranges of values as arguments which makes is really easy to learn the library.

Removed `prelude`, the library root can be directly included like `use holidays::*` now, it's simpler and does the same.

Added support for `time` library besides just `chrono`. Both are optional and only allow conversion to respective time representations.

### Internals

I added a `Date` type that's convertible to both std `SystemTime` and external types. User can't access this type. It was needed as a phf compatible alternative to `chrono` types - it's also way simpler and only needs to keep track of dates (not time).

## Dependencies

I removed hard dependency on `chrono`. It's an overkill given that this library only needs to track dates (not time), and _that_ is only used for querying data.<br/>
Instead I added a features that allows `Query` to seamlessly integrate with `chrono` types.

Replaced `once_cell` with `phf`.

# TODO

- [x] Finish `DateExt`
- [x] Add tests
- [x] Cleanup code
- [x] Add docs
- ~~Provide an error when user attempts to query dates outside of generated data - to allow differentiating "no holidays" and "no data available". Unbounded ranges already support unbounded querying.~~
  - Decided against it, would be annoying when querying multiple countries. Provided a function to query bounds instead.