@Caellian 2025-07-04
Closing in favor of #18.

I decided to do another larger rework and needed the `main` branch. I'll probably write a new crate that requires 2 additional crates in project structure so I split the work so far into a different branch so you can merge if you want.